+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.50.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.50.0)
+_March 12, 2024_
+
 ### ✅ Added
 - Link detection in the text views
 - Indicator when a message was edited